refactor(core): split the read stack into view, session, cache, and grep modules
