+++ crates/loonfs-core/src/path/read/grep.rs
@@ -1,23 +1,31 @@
-//! Content-search machinery: gram posting reads over index segments, the
-//! exhaustive-scan tail, and line-oriented match extraction. The
-//! orchestrating `grep` read lives on `LoadedMetadataView`
-//! (`materialized_view.rs`); this module holds the pieces that need no
-//! access to the view's internals.
+//! Content search, whole: the orchestrating `grep` read on
+//! [`LoadedMetadataView`](super::materialized_view::LoadedMetadataView) and
+//! its machinery — gram posting reads over index segments, the
+//! exhaustive-scan tail, and line-oriented match extraction.
 
 use crate::checkpoint::{
     index_segment_corrupt, load_index_segment_data_block, load_index_segment_filter_block,
     load_index_segment_index_block, string_prefix_upper_bound, CacheAdmission, MetadataTableCache,
 };
-use crate::error::{CoreError, Result};
+use crate::error::{CoreError, MetadataProjectionLoadError, MetadataViewError, Result};
+use crate::metadata::RevisionRecord;
+use crate::path::helpers::parse_absolute_path_for_core;
 use crate::query::grep_plan::GramQueryPlan;
+use crate::storage::content::read_durable_content_bytes;
 use crate::wal::{load_validated_wal_chain, WalChainLoadRequest};
-use futures::future::try_join_all;
+use futures::future::{join_all, try_join_all};
 use loonfs_api::wire::control::HeadState;
-use loonfs_api::wire::index_grams::{lookup, Gram, IndexRow, INDEX_FAMILY_GRAMS};
-use loonfs_api::wire::manifest::{hex_decode_bytes, IndexFileRef};
+use loonfs_api::wire::index_grams::{
+    lookup, Gram, IndexGramsFeature, IndexRow, INDEX_FAMILY_GRAMS, INDEX_GRAMS_FEATURE_KEY,
+    INDEX_GRAMS_MAX_FILE_BYTES,
+};
+use loonfs_api::wire::manifest::{hex_decode_bytes, IndexFileRef, MetadataTableFamily};
 use loonfs_api::wire::sst_blocks::{decode_filter_block, index_blocks_for_key_range};
 use loonfs_api::wire::wal::WalDelta;
-use loonfs_api::{ChangeSeq, ContentRef, InodeId, NamespaceId, RevisionNo};
+use loonfs_api::{
+    decode_grep_cursor, encode_grep_cursor, ChangeSeq, ContentRef, GrepMatch, GrepPageCursor,
+    GrepRequest, GrepResponse, InodeId, NamespaceId, RevisionNo,
+};
 use loonfs_objectstore::ObjectStore;
 use std::collections::{BTreeMap, BTreeSet};
 
@@ -466,6 +474,453 @@ pub(super) fn fold_rejected_frontier(
 pub(super) struct VisiblePathChain {
     pub(super) path: String,
     pub(super) ancestors: Vec<InodeId>,
+}
+
+/// One grep candidate that survived the cheap visibility checks and awaits
+/// its content read, carrying everything match emission needs so the
+/// fetched batch is processed without further metadata lookups.
+struct GrepContentCandidate {
+    inode_id: InodeId,
+    revision: RevisionRecord,
+    path: String,
+    /// The declared content size exceeds the index eligibility cap, so no
+    /// read is scheduled: the file could never pass the post-read text
+    /// check, and the walk skips it as fully scanned.
+    oversized: bool,
+}
+
+impl<S: ObjectStore + ?Sized> super::materialized_view::LoadedMetadataView<'_, S> {
+    /// Content search over the view: index-accelerated candidates through
+    /// the `index.grams` watermark, an exhaustive scan of the unindexed
+    /// tail, and real-pattern verification of every candidate. Matches
+    /// order by `(inode_id, byte_offset)`. Two budgets bound a page — the
+    /// match limit and a verified-candidate budget — and the cursor
+    /// resumes strictly after the last candidate the page finished
+    /// scanning, bound to the request that issued it. Each page is
+    /// evaluated against the view it runs on and reports that head in
+    /// `head_seq`.
+    #[tracing::instrument(
+        level = "info",
+        name = "loon.phase",
+        err,
+        skip_all,
+        fields(phase = "grep")
+    )]
+    pub(crate) async fn grep(&self, store: &S, request: &GrepRequest) -> Result<GrepResponse> {
+        use crate::query::grep_plan::{plan_pattern, GramPlanOutcome};
+
+        let limit = match request.limit {
+            None => DEFAULT_GREP_PAGE_LIMIT,
+            Some(0) => {
+                return Err(CoreError::InvalidQuery(
+                    "limit must be greater than zero".to_owned(),
+                ));
+            }
+            Some(limit) if limit as usize > MAX_GREP_PAGE_LIMIT => {
+                return Err(CoreError::InvalidQuery(format!(
+                    "limit {limit} exceeds the maximum of {MAX_GREP_PAGE_LIMIT}"
+                )));
+            }
+            Some(limit) => limit as usize,
+        };
+        let fingerprint = request.fingerprint();
+        let resume = match &request.cursor {
+            Some(cursor) => {
+                let cursor = decode_grep_cursor(cursor)
+                    .map_err(|error| CoreError::InvalidCursor(error.to_string()))?;
+                if cursor.fingerprint != fingerprint {
+                    return Err(CoreError::InvalidCursor(
+                        "the cursor was issued by a different request; replaying it \
+                         under new criteria would silently skip results"
+                            .to_owned(),
+                    ));
+                }
+                if cursor.head_seq > self.head.seq {
+                    return Err(MetadataViewError::SnapshotUnavailable {
+                        requested_seq: cursor.head_seq,
+                        head_seq: self.head.seq,
+                    }
+                    .into());
+                }
+                Some((cursor.last_inode_id, cursor.last_byte_offset))
+            }
+            None => None,
+        };
+
+        let manifest = self.tables.manifest();
+        let Some(value) = manifest.payload.features.get(INDEX_GRAMS_FEATURE_KEY) else {
+            return Err(CoreError::FeatureNotMaterialized {
+                feature: INDEX_GRAMS_FEATURE_KEY.to_owned(),
+            });
+        };
+        let feature = IndexGramsFeature::from_value(value).map_err(|error| {
+            use loonfs_api::wire::index_grams::IndexGramsCodecError;
+            match error {
+                IndexGramsCodecError::UnsupportedFeatureVersion { .. } => {
+                    CoreError::FeatureNotMaterialized {
+                        feature: INDEX_GRAMS_FEATURE_KEY.to_owned(),
+                    }
+                }
+                error => CoreError::NamespaceCorrupt(format!(
+                    "namespace `{}` carries an unreadable index.grams feature value: {error}",
+                    self.namespace_id.as_str()
+                )),
+            }
+        })?;
+        if !feature.is_materialized() {
+            return Err(CoreError::FeatureNotMaterialized {
+                feature: INDEX_GRAMS_FEATURE_KEY.to_owned(),
+            });
+        }
+
+        // Line-anchored semantics: `^` and `$` match line boundaries, the
+        // grep-family contract. The planner parses with the same flags so
+        // its gram analysis sees the pattern the verifier runs.
+        let pattern = regex::bytes::RegexBuilder::new(&request.pattern)
+            .case_insensitive(request.case_insensitive)
+            .multi_line(true)
+            .build()
+            .map_err(|error| CoreError::InvalidQuery(error.to_string()))?;
+
+        // One view session serves the whole page: candidates in the same
+        // directory tree share ancestor bindings, tombstone checks, and
+        // path verifications, so the per-candidate metadata walks below hit
+        // the session caches instead of re-fetching per candidate.
+        let mut session = self.metadata_view().session();
+        // The scope filter tests durable identity: resolve the prefix to
+        // its inode once, then require it among each candidate's ancestors,
+        // so name-policy folding and path normalization apply exactly as
+        // they do to every other read. A prefix that resolves to nothing
+        // has no matches.
+        let scope_root = match &request.path_prefix {
+            Some(prefix) => {
+                let parsed = parse_absolute_path_for_core(prefix)?;
+                match session.resolve_visible_path(&parsed, false).await {
+                    Ok(resolved) => Some(resolved.inode_id),
+                    Err(error) if error.code() == loonfs_api::ErrorCode::PathNotFound => {
+                        return Ok(GrepResponse {
+                            namespace_id: self.namespace_id.clone(),
+                            head_seq: self.head.seq,
+                            built_through_seq: feature.built_through_seq,
+                            tail_scanned: true,
+                            matches: Vec::new(),
+                            next_cursor: None,
+                        });
+                    }
+                    Err(error) => return Err(error),
+                }
+            }
+            None => None,
+        };
+
+        let mut candidates = GrepCandidates::default();
+        match plan_pattern(&request.pattern, request.case_insensitive)
+            .map_err(CoreError::InvalidQuery)?
+        {
+            GramPlanOutcome::Indexable(plan) => {
+                candidates.indexed = indexed_candidates(
+                    store,
+                    self.tables.table_cache(),
+                    &manifest.payload.index_files,
+                    &plan,
+                )
+                .await?;
+            }
+            GramPlanOutcome::Unindexable => {
+                if !request.allow_scan {
+                    return Err(CoreError::QueryUnindexable(
+                        "the pattern has no run of at least 3 literal bytes for the \
+                         trigram index; set allow_scan to search without it"
+                            .to_owned(),
+                    ));
+                }
+                candidates.unfiltered = self.scan_candidate_inodes().await?;
+            }
+        }
+
+        let tail = tail_revisions(
+            store,
+            &self.namespace_id,
+            &self.head,
+            manifest.payload.retention_floor_seq,
+            feature.built_through_seq,
+        )
+        .await?;
+        let mut tail_scanned = true;
+        if tail.len() > MAX_GREP_TAIL_FILES {
+            if request.allow_stale {
+                tail_scanned = false;
+            } else {
+                return Err(CoreError::IndexLagging {
+                    behind_commits: self.head.seq.0.saturating_sub(feature.built_through_seq.0),
+                });
+            }
+        } else {
+            candidates.unfiltered.extend(tail.keys().copied());
+        }
+
+        let mut matches: Vec<GrepMatch> = Vec::new();
+        let mut verified_files = 0usize;
+        let mut examined_candidates = 0usize;
+        let mut has_more = false;
+        // Where the next page resumes when this one stops early: the last
+        // candidate this page finished scanning (offset MAX), or the last
+        // emitted match when the page filled mid-file.
+        let mut resume_cursor: Option<(InodeId, u64)> = None;
+        // Highest candidate this page examined and rejected (invisible,
+        // superseded, or out of scope). A rejection is final for the page,
+        // so budget exits fold this into the cursor — without it, a run of
+        // rejections longer than a page budget would resume at the same
+        // cursor and re-reject the same candidates forever.
+        let mut rejected_frontier: Option<InodeId> = None;
+        let ordered_candidates = candidates.inodes().collect::<Vec<_>>();
+        let mut next_candidate = 0usize;
+        'page: loop {
+            // Select the next fan-out batch: walk candidates in inode
+            // order through the cheap checks (metadata lookups served by
+            // the loaded view) until enough survivors need content, the
+            // verified-file budget fills, or the candidates run out. The
+            // content read is the only per-candidate store fetch, so it is
+            // the only stage that fans out — the design doc's "small fixed
+            // fan-out" for candidate reads.
+            let mut batch: Vec<GrepContentCandidate> = Vec::new();
+            let mut budget_exhausted = false;
+            while next_candidate < ordered_candidates.len() {
+                let inode_id = ordered_candidates[next_candidate];
+                if let Some((last_inode, last_offset)) = resume {
+                    if inode_id < last_inode || (inode_id == last_inode && last_offset == u64::MAX)
+                    {
+                        next_candidate += 1;
+                        continue;
+                    }
+                }
+                // The examination budget bounds a page's metadata work the
+                // way the verified budget bounds its content work: a scope
+                // filter that rejects nearly every candidate would
+                // otherwise walk metadata for the entire candidate set in
+                // one page. The candidate at the boundary is left for the
+                // next page.
+                if examined_candidates == MAX_GREP_EXAMINED_CANDIDATES_PER_PAGE {
+                    budget_exhausted = true;
+                    break;
+                }
+                next_candidate += 1;
+                examined_candidates += 1;
+                if session.visible_inode(inode_id).await?.is_none() {
+                    rejected_frontier = Some(inode_id);
+                    continue;
+                }
+                let Some(revision) = session.latest_revision_head_of_visible(inode_id).await?
+                else {
+                    rejected_frontier = Some(inode_id);
+                    continue;
+                };
+                if !candidates.admits(inode_id, revision.revision_no) {
+                    rejected_frontier = Some(inode_id);
+                    continue;
+                }
+                // With the tail skipped (`allow_stale`), serve the index's
+                // cut and nothing newer: a candidate whose newest revision
+                // is past the watermark would otherwise be verified at an
+                // unindexed revision while tail-only files stay invisible
+                // — a mix of two snapshots rather than a stale-but-
+                // consistent one.
+                if !tail_scanned && revision.committed_seq > feature.built_through_seq {
+                    rejected_frontier = Some(inode_id);
+                    continue;
+                }
+                let Some(chain) = derive_visible_path(&mut session, inode_id).await? else {
+                    rejected_frontier = Some(inode_id);
+                    continue;
+                };
+                if let Some(scope_root) = scope_root {
+                    if !chain.ancestors.contains(&scope_root) {
+                        rejected_frontier = Some(inode_id);
+                        continue;
+                    }
+                }
+                if verified_files == MAX_GREP_VERIFIED_FILES_PER_PAGE {
+                    budget_exhausted = true;
+                    break;
+                }
+                verified_files += 1;
+                // Content past the index eligibility cap can never be
+                // indexable text, so tail and scan candidates skip their
+                // doomed reads on the declared size alone — the same
+                // pre-fetch check the index builder applies
+                // (`checkpoint/index_build.rs`); index-supplied candidates
+                // are under the cap by construction. The candidate still
+                // rides the batch in inode order, so the budget and the
+                // resume cursor advance exactly as if its bytes had been
+                // fetched and refused.
+                let oversized = revision.content_ref.size_bytes > INDEX_GRAMS_MAX_FILE_BYTES;
+                batch.push(GrepContentCandidate {
+                    inode_id,
+                    revision,
+                    path: chain.path,
+                    oversized,
+                });
+                if batch.len() == MAX_GREP_CONTENT_IO {
+                    break;
+                }
+            }
+            if batch.is_empty() {
+                if budget_exhausted {
+                    has_more = true;
+                    fold_rejected_frontier(&mut resume_cursor, rejected_frontier);
+                }
+                break 'page;
+            }
+            // The reads fan out, but their errors do not short-circuit:
+            // each result rides with its candidate into the ordered walk
+            // below, which surfaces a failure only when it reaches that
+            // candidate — the position the serial loop surfaced it. A
+            // failure the walk never reaches (the page filled first) is
+            // discarded with the rest of the speculative batch; the next
+            // page re-issues that read and reports it then. An oversized
+            // candidate carries no read at all (`None`).
+            let contents = join_all(batch.iter().map(|candidate| async move {
+                if candidate.oversized {
+                    return None;
+                }
+                Some(
+                    read_durable_content_bytes(
+                        store,
+                        &self.content_store_id,
+                        &candidate.revision.content_ref,
+                    )
+                    .await,
+                )
+            }))
+            .await;
+            // Emission stays strictly in candidate (inode) order: the batch
+            // was selected in order and its results are consumed in order,
+            // so matches, limits, errors, and the resume cursor advance
+            // exactly as the serial walk advanced them.
+            for (candidate, content) in batch.iter().zip(contents) {
+                let inode_id = candidate.inode_id;
+                let Some(content) = content else {
+                    // Skipped as oversized: scanned-and-refused without the
+                    // fetch, so the cursor moves past it like any other
+                    // ineligible file.
+                    resume_cursor = Some((inode_id, u64::MAX));
+                    continue;
+                };
+                let content = content?;
+                if !crate::checkpoint::is_indexable_text_content(&content.bytes) {
+                    resume_cursor = Some((inode_id, u64::MAX));
+                    continue;
+                }
+                for found in line_matches(&content.bytes, &pattern) {
+                    if let Some((last_inode, last_offset)) = resume {
+                        if inode_id == last_inode && found.byte_offset <= last_offset {
+                            continue;
+                        }
+                    }
+                    if matches.len() == limit {
+                        // The page filled mid-batch: contents already
+                        // fetched for the batch's later candidates are
+                        // discarded, and the cursor resumes from the last
+                        // processed candidate, never a discarded one.
+                        has_more = true;
+                        break 'page;
+                    }
+                    resume_cursor = Some((inode_id, found.byte_offset));
+                    matches.push(GrepMatch {
+                        absolute_path: candidate.path.clone(),
+                        inode_id,
+                        revision_no: candidate.revision.revision_no,
+                        line_number: found.line_number,
+                        byte_offset: found.byte_offset,
+                        line: found.line,
+                        line_truncated: found.line_truncated,
+                    });
+                }
+                // The file was fully scanned; a later stop resumes past it.
+                resume_cursor = Some((inode_id, u64::MAX));
+            }
+            if budget_exhausted {
+                has_more = true;
+                // The whole final batch was scanned, so every examined
+                // candidate is resolved; rejections past the last scanned
+                // file move the cursor with them.
+                fold_rejected_frontier(&mut resume_cursor, rejected_frontier);
+                break 'page;
+            }
+        }
+
+        let next_cursor = if has_more {
+            let (last_inode_id, last_byte_offset) = resume_cursor.or(resume).ok_or_else(|| {
+                CoreError::Internal(
+                    "a truncated page must have scanned at least one candidate".to_owned(),
+                )
+            })?;
+            Some(
+                encode_grep_cursor(&GrepPageCursor {
+                    head_seq: self.head.seq,
+                    last_inode_id,
+                    last_byte_offset,
+                    fingerprint,
+                })
+                .map_err(|error| CoreError::Internal(error.to_string()))?,
+            )
+        } else {
+            None
+        };
+        Ok(GrepResponse {
+            namespace_id: self.namespace_id.clone(),
+            head_seq: self.head.seq,
+            built_through_seq: feature.built_through_seq,
+            tail_scanned,
+            matches,
+            next_cursor,
+        })
+    }
+
+    /// Every inode holding any revision in the manifest tables, for
+    /// plan-less scans; the WAL tail is collected separately. Refuses past
+    /// the scan budget.
+    async fn scan_candidate_inodes(&self) -> Result<std::collections::BTreeSet<InodeId>> {
+        let mut inodes = std::collections::BTreeSet::new();
+        let mut lower = "revision-".to_owned();
+        let upper = crate::checkpoint::string_prefix_upper_bound("revision-");
+        loop {
+            let page = self
+                .tables
+                .scan_range_page_with_keys(
+                    MetadataTableFamily::Revisions,
+                    &lower,
+                    upper.as_deref(),
+                    MAX_GREP_SCAN_FILES.saturating_add(1),
+                )
+                .await
+                .map_err(|error| {
+                    CoreError::MetadataProjection(MetadataProjectionLoadError::ManifestLoad(error))
+                })?;
+            let Some((last_key, _)) = page.last() else {
+                break;
+            };
+            let last_key = last_key.clone();
+            let exhausted = page.len() <= MAX_GREP_SCAN_FILES;
+            for (_, row) in page {
+                if let loonfs_api::wire::manifest::MetadataRow::Revision { inode_id, .. } = row {
+                    inodes.insert(inode_id);
+                }
+            }
+            if inodes.len() > MAX_GREP_SCAN_FILES {
+                return Err(CoreError::QueryUnindexable(format!(
+                    "the namespace exceeds the {MAX_GREP_SCAN_FILES}-file scan budget; \
+                     give the pattern a run of at least 3 literal bytes so the \
+                     trigram index can narrow candidates"
+                )));
+            }
+            if exhausted {
+                break;
+            }
+            lower = format!("{last_key}\0");
+        }
+        Ok(inodes)
+    }
 }
 
 #[cfg(test)]
